terraform: redo how flattening works
