+++ terraform/graph_config_node_module.go
@@ -166,22 +166,6 @@ func (n *graphNodeModuleExpanded) FlattenGraph() *Graph {
 	graph := n.Subgraph()
 	input := n.Original.Module.RawConfig
 
-	// Build the string that represents the path. We do this once here
-	// so that we only have to compute it once. The block below is in {}
-	// so that parts drops out of scope immediately.
-	var pathStr, parentPathStr string
-	{
-		parts := make([]string, 0, len(graph.Path)*2)
-		for _, p := range graph.Path[1:] {
-			parts = append(parts, "module", p)
-		}
-
-		pathStr = strings.Join(parts, ".")
-		if len(parts) > 2 {
-			parentPathStr = strings.Join(parts[0:len(parts)-2], ".")
-		}
-	}
-
 	// Go over each vertex in the graph and wrap the configuration
 	// items so that the dependencies properly map to the modules.
 	// See the docs for graphNodeModuleWrappable for more info.
@@ -190,15 +174,6 @@ func (n *graphNodeModuleExpanded) FlattenGraph() *Graph {
 			graph.Remove(v)
 			continue
 		}
-
-		wn, ok := v.(graphNodeModuleWrappable)
-		if !ok {
-			panic("unwrappable node: " + dag.VertexName(v))
-		}
-
-		// Prefix for dependencies. We set this to blank for variables
-		// (see below).
-		depPrefix := pathStr
 
 		// If this is a variable, then look it up in the raw configuration.
 		// If it exists in the raw configuration, set the value of it.
@@ -217,20 +192,7 @@ func (n *graphNodeModuleExpanded) FlattenGraph() *Graph {
 				// Set the variable value so it is interpolated properly
 				vn.SetVariableValue(config)
 			}
-
-			// We set the dependency prefix to the parent path
-			// since variables can reference outside of our own graph
-			// to the parent graph.
-			depPrefix = parentPathStr
 		}
-
-		graph.Replace(v, &graphNodeModuleFlatWrap{
-			graphNodeModuleWrappable: wn,
-
-			PathValue:         graph.Path,
-			NamePrefix:        pathStr,
-			DependentOnPrefix: depPrefix,
-		})
 	}
 
 	return graph
@@ -247,58 +209,57 @@ type graphNodeModuleSkippable interface {
 	FlattenSkip() bool
 }
 
-// This is the interface that all nodes within a module graph must
-// implement to be wrapped for flattening.
-type graphNodeModuleWrappable interface {
-	graphNodeConfig
-}
-
+/*
 // graphNodeModuleFlatWrap wraps elements of the module graph
 // when they are flattened so that the dependency information is
 // correct.
 type graphNodeModuleFlatWrap struct {
-	graphNodeModuleWrappable
-
+	Vertex            dag.Vertex
 	PathValue         []string
 	NamePrefix        string
 	DependentOnPrefix string
 }
 
-func (n *graphNodeModuleFlatWrap) Name() string {
-	return fmt.Sprintf("%s.%s", n.NamePrefix, n.graphNodeModuleWrappable.Name())
-}
-
-// GraphNodeSubPath impl.
-func (n *graphNodeModuleFlatWrap) Path() []string {
-	return n.PathValue
-}
-
-func (n *graphNodeModuleFlatWrap) DependableName() []string {
-	result := n.graphNodeModuleWrappable.DependableName()
-	n.prefixList(result, n.NamePrefix)
-	return result
-}
-
-func (n *graphNodeModuleFlatWrap) DependentOn() []string {
-	result := n.graphNodeModuleWrappable.DependentOn()
-	if n.DependentOnPrefix != "" {
-		n.prefixList(result, n.DependentOnPrefix)
-	}
-
-	return result
-}
-
-func (n *graphNodeModuleFlatWrap) Proxy() bool {
-	pn, ok := n.graphNodeModuleWrappable.(GraphNodeProxy)
+// GraphNodeProvider impl.
+func (n *graphNodeModuleFlatWrap) ProviderName() string {
+	pn, ok := n.Vertex.(GraphNodeProvider)
 	if !ok {
-		return false
+		return ""
 	}
 
-	return pn.Proxy()
+	return fmt.Sprintf("%s.%s", n.NamePrefix, pn.ProviderName())
 }
 
-func (n *graphNodeModuleFlatWrap) prefixList(result []string, prefix string) {
-	for i, v := range result {
-		result[i] = fmt.Sprintf("%s.%s", prefix, v)
+// GraphNodeProviderConsumer impl.
+func (n *graphNodeModuleFlatWrap) ProvidedBy() []string {
+	pn, ok := n.Vertex.(GraphNodeProviderConsumer)
+	if !ok {
+		return nil
 	}
+
+	result := make([]string, 0, 2)
+	for _, v := range pn.ProvidedBy() {
+		result = append(result, fmt.Sprintf("%s.%s", n.NamePrefix, v))
+	}
+	return result
+}
+*/
+
+func modulePrefixStr(p []string) string {
+	parts := make([]string, 0, len(p)*2)
+	for _, p := range p[1:] {
+		parts = append(parts, "module", p)
+	}
+
+	return strings.Join(parts, ".")
+}
+
+func modulePrefixList(result []string, prefix string) []string {
+	if prefix != "" {
+		for i, v := range result {
+			result[i] = fmt.Sprintf("%s.%s", prefix, v)
+		}
+	}
+
+	return result
 }
