+++ terraform/graph_config_node_module_test.go
@@ -1,7 +1,6 @@
 package terraform
 
 import (
-	"reflect"
 	"strings"
 	"testing"
 
@@ -59,80 +58,13 @@ func TestGraphNodeConfigModuleExpandFlatten(t *testing.T) {
 		t.Fatalf("err: %s", err)
 	}
 
-	fg := g.(GraphNodeFlattenable)
+	fg := g.(GraphNodeFlatGraph)
 
 	actual := strings.TrimSpace(fg.FlattenGraph().String())
 	expected := strings.TrimSpace(testGraphNodeModuleExpandFlattenStr)
 	if actual != expected {
 		t.Fatalf("bad:\n\n%s", actual)
 	}
-}
-
-func TestGraphNodeModulFlatWrap_Name(t *testing.T) {
-	n := &graphNodeModuleFlatWrap{
-		graphNodeModuleWrappable: &testGraphNodeModuleWrappable{
-			NameValue: "foo",
-		},
-
-		NamePrefix: "module.bar",
-	}
-
-	if v := n.Name(); v != "module.bar.foo" {
-		t.Fatalf("bad: %s", v)
-	}
-}
-
-func TestGraphNodeModulFlatWrap_DependentOn(t *testing.T) {
-	n := &graphNodeModuleFlatWrap{
-		graphNodeModuleWrappable: &testGraphNodeModuleWrappable{
-			NameValue: "foo",
-		},
-
-		NamePrefix:        "module.bar",
-		DependentOnPrefix: "module.bar",
-	}
-
-	actual := n.DependentOn()
-	expected := []string{"module.bar.foo"}
-	if !reflect.DeepEqual(actual, expected) {
-		t.Fatalf("bad: %#v", actual)
-	}
-}
-
-func TestGraphNodeModulFlatWrap_DependableName(t *testing.T) {
-	n := &graphNodeModuleFlatWrap{
-		graphNodeModuleWrappable: &testGraphNodeModuleWrappable{
-			NameValue: "foo",
-		},
-
-		NamePrefix: "module.bar",
-	}
-
-	actual := n.DependableName()
-	expected := []string{"module.bar.foo"}
-	if !reflect.DeepEqual(actual, expected) {
-		t.Fatalf("bad: %#v", actual)
-	}
-}
-
-type testGraphNodeModuleWrappable struct {
-	NameValue string
-}
-
-func (n *testGraphNodeModuleWrappable) ConfigType() GraphNodeConfigType {
-	return GraphNodeConfigTypeInvalid
-}
-
-func (n *testGraphNodeModuleWrappable) Name() string {
-	return n.NameValue
-}
-
-func (n *testGraphNodeModuleWrappable) DependableName() []string {
-	return []string{"foo"}
-}
-
-func (n *testGraphNodeModuleWrappable) DependentOn() []string {
-	return []string{"foo"}
 }
 
 const testGraphNodeModuleExpandStr = `
@@ -144,5 +76,5 @@ module inputs
 `
 
 const testGraphNodeModuleExpandFlattenStr = `
-module.child.aws_instance.foo
+aws_instance.foo
 `
