+++ terraform/graph_config_node_test.go
@@ -108,11 +108,3 @@ func TestGraphNodeConfigResource_ProvisionedBy(t *testing.T) {
 		t.Fatalf("bad: %#v", actual)
 	}
 }
-
-func TestGraphNodeConfigVariable_impl(t *testing.T) {
-	var _ dag.Vertex = new(GraphNodeConfigVariable)
-	var _ dag.NamedVertex = new(GraphNodeConfigVariable)
-	var _ graphNodeConfig = new(GraphNodeConfigVariable)
-	var _ GraphNodeVariable = new(GraphNodeConfigVariable)
-	var _ GraphNodeProxy = new(GraphNodeConfigVariable)
-}
